[CI] finalize dv setup for all passing verification
- Failing ones are added to`allow_failures`

- After clearing build.out, the pass/fail values are now reflective of the actual number.

I believe the next step after figuring out what's wrong with the pll TB and the other wb_utests is to move the heavy CI tests to a new branch. This should follow a similar concept of staging in openlane, but will not be a gate between develop and staging and will remain out of sync to only be updated within the context of the CI. This new branch should carry the DV and full LVS tests + whatever comes to mind later, while develop will keep the base checks, the precheck, and the abstract lvs.